+++ src/common/sourceMaps/renameWorker.ts
@@ -2,23 +2,19 @@
  * Copyright (C) Microsoft Corporation. All rights reserved.
  *--------------------------------------------------------*/
 
-import { Node, SourceLocation } from 'estree';
+import { Node } from 'estree';
 import { isMainThread, parentPort, workerData } from 'worker_threads';
 import { parseProgram, traverse } from '../sourceCodeManipulations';
 import type { FlatTree } from './renameScopeTree';
-
-type TypeWithChildren = { children?: TypeWithChildren[] };
 
 /**
  * Gets ranges of scopes in the source code. It returns ranges where variables
  * declared in those ranges apply to all child scopes. For example,
  * `function(foo) { bar(); }` emits `(foo) { bar(); }` as a range.
  */
-export function extractScopeRangesWithFactory<T extends TypeWithChildren>(
-  source: string,
-  factory: (start: SourceLocation, end: SourceLocation) => T,
-) {
+export function extractScopeRangesWithFactory(source: string): FlatTree {
   const program = parseProgram(source);
+  const output: FlatTree = [];
 
   const push = (
     indexingNode: Node,
@@ -29,22 +25,14 @@ export function extractScopeRangesWithFactory<T extends TypeWithChildren>(
       throw new Error('should include locations');
     }
 
-    const next = factory(start, end);
-    if (stack.length > 0) {
-      const parent = stack[stack.length - 1];
-      parent.scopeNode.children ??= [];
-      parent.scopeNode.children.push(next);
-    }
-
-    stack.push({ node: indexingNode, scopeNode: next });
+    output.push({ start: start.start, end: end.end, depth: stack.length });
+    stack.push(indexingNode);
   };
 
   // nodes ignored because they're already captured in the top
   // level statement, such as the body of `for` loops.
   const coveredBlocks = new Set<Node>();
-
-  const stack: { node: Node; scopeNode: T }[] = [];
-  let programScope: T | undefined;
+  const stack: Node[] = [];
 
   traverse(program, {
     enter: node => {
@@ -58,7 +46,6 @@ export function extractScopeRangesWithFactory<T extends TypeWithChildren>(
         // include the top level program:
         case 'Program':
           push(node);
-          programScope = stack[0].scopeNode;
           break;
         // include the entire loop to handle declarations inside the statement:
         case 'ForStatement':
@@ -76,20 +63,15 @@ export function extractScopeRangesWithFactory<T extends TypeWithChildren>(
       }
     },
     leave: node => {
-      if (node === stack[stack.length - 1].node) {
+      if (node === stack[stack.length - 1]) {
         stack.pop();
       }
     },
   });
 
-  return programScope as T;
+  return output;
 }
 
 if (!isMainThread) {
-  parentPort?.postMessage(
-    extractScopeRangesWithFactory<FlatTree>(workerData, (start, end) => ({
-      start: start.start,
-      end: end.end,
-    })),
-  );
+  parentPort?.postMessage(extractScopeRangesWithFactory(workerData));
 }